Added pop-up for 'Delete Answer'

+++ resources/views/answer.blade.php
@@ -10,7 +10,8 @@
                         {{$answer->body}}
                     </div>
                     <div class="card-footer">
-                        {{ Form::open(['method'  => 'DELETE', 'route' => ['answer.destroy', $question, $answer->id]])}}
+                        {{ Form::open(['method'  => 'DELETE','onsubmit'=> "return confirm('Do you really want to delete?');",'route' => ['answer.destroy', $question, $answer->id]])}}
+                        <input type="hidden" name="_method" value="DELETE" />
                         <button class="btn btn-danger float-right mr-2" value="submit" type="submit" id="submit">Delete
                         </button>
                         {!! Form::close() !!}
